adding ResponseEntity and Exceptions handling;

+++ src/main/java/com/mine/jwt/Controller/User/UserController.java
@@ -3,6 +3,7 @@ package com.mine.jwt.Controller.User;
 import com.mine.jwt.Models.Domain.Users.User;
 import com.mine.jwt.Repository.User.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -25,18 +26,21 @@ public class UserController {
     }
 
     @GetMapping(value = "/users")
-    public List<User> getAll() {
-        return userRepo.findAll();
+    public ResponseEntity<List<User>> getAll() {
+        return ResponseEntity.ok(userRepo.findAll());
     }
 
     @GetMapping(value = "/users/{id}")
-    public User getSingle(@PathVariable String id) {
-        return userRepo.findById(id).get();
+    public ResponseEntity<User> getSingle(@PathVariable String id) throws RuntimeException {
+        if (userRepo.findById(id).isEmpty()) throw new RuntimeException("user not found");
+        return ResponseEntity.ok(userRepo.findById(id).get());
     }
 
     @DeleteMapping(value = "/users/delete/{id}")
-    public void del(@PathVariable String id) {
+    public ResponseEntity<String> del(@PathVariable String id) throws RuntimeException {
+        if (userRepo.findById(id).isEmpty()) throw new RuntimeException("user not found");
         userRepo.deleteById(id);
+        return ResponseEntity.status(202).body("successfully deleted");
     }
 }
